refactor(editor): split DetailPanel into kind-specific components
- DetailPanel.svelte is now a thin Dialog shell that switches content
  by element kind (node, edge, subgraph, port)
- NodeDetail: icon + label + spec binding + connections + PoE
- EdgeDetail: from/to endpoints + bandwidth/VLAN/label properties
- SubgraphDetail: label + children/direction/bounds info
- PortDetail: label/side + connection lookup from links
- Remove tabs (Overview/Properties/Raw) — each kind shows relevant info
